+++ app/src/main/java/com/rygital/moneytracker/utils/StringFormatter.kt
@@ -1,3 +1,3 @@
 package com.rygital.moneytracker.utils
 
-fun formatMoney(value: Float) = String.format("%.2f", value)
+fun formatMoney(value: Double) = String.format("%.2f", value)
